+++ src/components/__tests__/modal.exercise.js
@@ -1,0 +1,11 @@
+// 🐨 you're gonna need this stuff:
+// import {Modal, ModalContents, ModalOpenButton} from '../modal'
+
+test.todo('can be opened and closed')
+// 🐨 render the Modal, ModalOpenButton, and ModalContents
+// 🐨 click the open button
+// 🐨 verify the modal contains the modal contents, title, and label
+// 🐨 click the close button
+// 🐨 verify the modal is no longer rendered
+// 💰 (use `query*` rather than `get*` or `find*` queries to verify it is not rendered)
+// 💰 Remember all userEvent utils are async, so you need to await them.
